Creating MeterProvider API
