+++ sdk/metrics/src/MeterProviderProxy.cpp
@@ -26,29 +26,4 @@ libmexclass::proxy::MakeResult MeterProviderProxy::make(const libmexclass::proxy
     return out;
 }
 
-void MeterProviderProxy::getMeter(libmexclass::proxy::method::Context& context) {
-   // Always assumes 3 inputs
-   matlab::data::StringArray name_mda = context.inputs[0];
-   std::string name = static_cast<std::string>(name_mda[0]);
-   matlab::data::StringArray version_mda = context.inputs[1];
-   std::string version = static_cast<std::string>(version_mda[0]);
-   matlab::data::StringArray schema_mda = context.inputs[2];
-   std::string schema = static_cast<std::string>(schema_mda[0]); 
-	
-   auto mt = CppMeterProvider->GetMeter(name, version, schema);
-
-   // instantiate a MeterProxy instance
-   MeterProxy* newproxy = new MeterProxy(mt);
-   auto mtproxy = std::shared_ptr<libmexclass::proxy::Proxy>(newproxy);
-    
-   // obtain a proxy ID
-   libmexclass::proxy::ID proxyid = libmexclass::proxy::ProxyManager::manageProxy(mtproxy);
-
-   // return the ID
-   matlab::data::ArrayFactory factory;
-   auto proxyid_mda = factory.createScalar<libmexclass::proxy::ID>(proxyid);
-   context.outputs[0] = proxyid_mda;
-}
-
-
 } // namespace libmexclass::opentelemetry
